initial break up
